+++ apps/src/componentLibrary/tooltip/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [0.2.1] (https://github.com/code-dot-org/code-dot-org/pull/59610)
 
-* used DSCO Button component for stories with Multiple Tooltips
+* used DSCO Button component for `Tooltip` stories
 
 ## [0.2.0](https://github.com/code-dot-org/code-dot-org/pull/59106)